Mapx dtype updates (#1010)
Refs #995 and #994
Adds a Jupyter Notebook script that compares the mapx files to the
database tables they reference and updates the dtypes of the mapx
accordingly. This commit also includes all of the mapx files that were
updated on the initial execution of the script.

It would be great to have this type of script be able to run
automatically as a CI/CD workflow whenever a PR is submitted. But for
now, I recommend that we add running this script to our deployment
checklist as another manual execution. The execution results in modified
mapx files, which should then be committed to a new branch, pushed, and
opened as a PR.